Rename missingViewBox file to be consistent with other icon svg file names
