+++ src/icons/SVGIcon/data/missing-view-box.js
